backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
fakeredis==2.18.1
geopy==2.4.0
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1